chore: factor out CompressorImpl into separate files
